+++ .gitignore
@@ -1,2 +1,3 @@
 .idea/
 data/
+config/config.py
